# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/images/bulk-update": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Bulk Update Images",
        "description": "Apply a bulk update to multiple images simultaneously.\n\nThis endpoint takes a list of image IDs and a data payload, applying the changes to all specified images in a single transaction. List-like fields (e.g. tags) can be appended or overwritten based on the operation_mode.",
        "operationId": "bulk_update_images_api_images_bulk_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageBulkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Bulk Update Images Api Images Bulk Update Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/bulk-move": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Bulk Move Images",
        "description": "Move multiple images to a different set.",
        "operationId": "bulk_move_images_api_images_bulk_move_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageBulkMove"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Bulk Move Images Api Images Bulk Move Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Images",
        "description": "Retrieve a paginated list of all images in the vault.\n\nSupports comprehensive filtering via `search` (matching filename, tags, or notes) and `rating`. Returns images enriched with context like their parent set title and associated creators.",
        "operationId": "read_images_api_images__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rating"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImagePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/duplicates/resolve": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Resolve Duplicates",
        "description": "Resolve a group of visually identical duplicate images.\n\nThis endpoint retains the specified `keep_image_id` and permanently deletes the files and database records for all `remove_image_ids`. Use this carefully as deletion is irreversible.",
        "operationId": "resolve_duplicates_api_images_duplicates_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Resolve Duplicates Api Images Duplicates Resolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/random/file/{ratio}/tags/{tags}/image.jpg": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Random Image File Path Tags",
        "description": "Get a random image file based on ratio and tags in the path (DisplayFusion compatible).",
        "operationId": "read_random_image_file_path_tags_api_images_random_file__ratio__tags__tags__image_jpg_get",
        "parameters": [
          {
            "name": "ratio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ratio"
            }
          },
          {
            "name": "tags",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/random/file/{ratio}/image.jpg": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Random Image File Path",
        "description": "Get a random image file based on ratio in the path (DisplayFusion compatible).",
        "operationId": "read_random_image_file_path_api_images_random_file__ratio__image_jpg_get",
        "parameters": [
          {
            "name": "ratio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ratio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/random": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Random Image",
        "description": "Get a random image based on filters.",
        "operationId": "read_random_image_api_images_random_get",
        "parameters": [
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "aspect_ratio_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aspect Ratio Label"
            }
          },
          {
            "name": "min_width",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Width"
            }
          },
          {
            "name": "min_height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Height"
            }
          },
          {
            "name": "creator_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/random/file": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Random Image File",
        "description": "Get a random image file based on filters.",
        "operationId": "read_random_image_file_api_images_random_file_get",
        "parameters": [
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "aspect_ratio_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aspect Ratio Label"
            }
          },
          {
            "name": "min_width",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Width"
            }
          },
          {
            "name": "min_height",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Height"
            }
          },
          {
            "name": "creator_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creator Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Read Image",
        "operationId": "read_image_api_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "images"
        ],
        "summary": "Update Image",
        "operationId": "update_image_api_images__image_id__patch",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "Delete Image",
        "operationId": "delete_image_api_images__image_id__delete",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/file/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Image File",
        "operationId": "get_image_file_api_images_file__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/images/set/{set_id}": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Create Image For Set",
        "operationId": "create_image_for_set_api_images_set__set_id__post",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkOperationMode": {
        "type": "string",
        "enum": [
          "replace",
          "append",
          "remove"
        ],
        "title": "BulkOperationMode"
      },
...
      "DuplicateResolutionRequest": {
        "properties": {
          "keep_image_id": {
            "type": "integer",
            "title": "Keep Image Id",
            "description": "The ID of the image to keep. All others will be deleted."
          },
          "remove_image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Remove Image Ids",
            "description": "List of image IDs to permanently delete from disk and database."
          }
        },
        "type": "object",
        "required": [
          "keep_image_id",
          "remove_image_ids"
        ],
        "title": "DuplicateResolutionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImageBulkMove": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Image Ids",
            "description": "List of image IDs to move."
          },
          "target_set_id": {
            "type": "integer",
            "title": "Target Set Id",
            "description": "The ID of the destination set."
          }
        },
        "type": "object",
        "required": [
          "image_ids",
          "target_set_id"
        ],
        "title": "ImageBulkMove"
      },
      "ImageBulkUpdate": {
        "properties": {
          "image_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Image Ids",
            "description": "List of image IDs to apply the bulk update to."
          },
          "update_data": {
            "$ref": "#/components/schemas/ImageUpdate",
            "description": "The data to apply to all selected images."
          },
          "operation_mode": {
            "$ref": "#/components/schemas/BulkOperationMode",
            "description": "How to apply list-like fields (e.g., tags). APPEND or OVERWRITE.",
            "default": "append"
          }
        },
        "type": "object",
        "required": [
          "image_ids",
          "update_data"
        ],
        "title": "ImageBulkUpdate"
      },
...
      "ImagePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImageWithContext"
            },
            "type": "array",
            "title": "Items",
            "description": "Paginated list of images."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of images matching the query."
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "Number of items skipped."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum number of items returned."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "ImagePage"
      },
...
    }
  }
}